% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bayesian Helper Functions.R
\name{betaBinomPosteriorPars}
\alias{betaBinomPosteriorPars}
\title{Function to calculate beta-binomial posterior parameters}
\usage{
betaBinomPosteriorPars(Ni = NULL, Si = NULL, Fi = NULL, Nk = NULL,
  Sk = NULL, Fk = NULL, equiv.obs = NULL, pars = NULL)
}
\arguments{
\item{Ni}{Number of tests/observations in the data (must be positive if supplied)}

\item{Si}{Number of successful tests/observations in the data (must be positive if supplied)}

\item{Fi}{Number of failed tests/observations in the data (must be positive if supplied)}

\item{Nk}{Number of tests/observations contributing to the prior distribution (must be positive if supplied)}

\item{Sk}{Number of successful tests/observations contributing to the prior distribution (must be positive if supplied)}

\item{Fk}{Number of failed tests/observations contributing to the prior distribution (must be positive if supplied)}

\item{equiv.obs}{Number of equivalent observations supplied by the prior distribution (must be positive if supplied)}

\item{pars}{prior distribution parameters shape1, shape2 which can be supplied instead of Nk, Sk, Fk, and eff.obs}
}
\value{
data frame with columns alpha and beta, representing parameters shape1 and shape2 of a beta distribution
}
\description{
Function to calculate beta-binomial posterior parameters
}
\details{
Two of Ni, Si, Fi must be supplied (non-null); the third will be calculated. Either equiv.obs and two of Nk, Sk, Fk must be supplied, or pars must be a vector of length two or a matrix of at least two columns.
Similarly, two of Nk, Sk, Fk must be supplied along with equiv.obs (the prior parameters will be calculated using this information)
}

