% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordCloud.R
\name{comparison.cloud}
\alias{comparison.cloud}
\title{Plot a cloud comparing the frequencies of words across documents.}
\usage{
comparison.cloud(term.matrix, scale = c(4, 0.5), max.words = 300,
  random.order = FALSE, rot.per = 0.1,
  colors = RColorBrewer::brewer.pal(ncol(term.matrix), "Dark2"),
  use.r.layout = FALSE, title.size = 3, ...)
}
\arguments{
\item{term.matrix}{A term frequency matrix whose rows represent words and whose columns represent documents.}

\item{scale}{A vector of length 2 indicating the range of the size of the words.}

\item{max.words}{Maximum number of words to be plotted. least frequent terms dropped}

\item{random.order}{plot words in random order. If false, they will be plotted in decreasing frequency}

\item{rot.per}{proportion words with 90 degree rotation}

\item{colors}{color words from least to most frequent}

\item{use.r.layout}{if false, then c++ code is used for collision detection, otherwise R is used}

\item{title.size}{Size of document titles}

\item{...}{Additional parameters to be passed to text (and strheight, strwidth).}
}
\description{
Plot a cloud comparing the frequencies of words across documents.
}

