% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bayesian Helper Functions.R
\name{equivBetaPars}
\alias{equivBetaPars}
\title{Function to calculate beta prior parameters with data from a more general population}
\usage{
equivBetaPars(Nk = NULL, Sk = NULL, Fk = NULL, equiv.obs = 5)
}
\arguments{
\item{Nk}{Number of trials/observations (must be positive if supplied)}

\item{Sk}{Number of successful trials (must be positive if supplied)}

\item{Fk}{Number of failed trials (must be positive if supplied)}

\item{equiv.obs}{Number of equivalent observations supplied by the prior distribution}
}
\value{
data frame with columns alpha and beta, representing parameters shape1 and shape2 of a beta distribution
}
\description{
Function to calculate beta prior parameters with data from a more general population
}
\details{
Two of Nk, Sk, Fk must be supplied (non-null); the third will be calculated.
}

