% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordCloud.R
\name{MakeWordcloud}
\alias{MakeWordcloud}
\title{Function to make a wordcloud}
\usage{
MakeWordcloud(x, color.set = RColorBrewer::brewer.pal(6, "Dark2"),
  colors = color.set, max.words = 50, rot.per = 0.3, random.order = F,
  random.color = T, ...)
}
\arguments{
\item{x}{word frequency data frame (from MakeWordFreq), with columns "word" and "freq"}

\item{color.set}{vector of colors}

\item{max.words}{maximum words}

\item{rot.per}{percentage of words to rotate}

\item{random.order}{plot words in random order?}

\item{random.color}{color words randomly?}

\item{...}{arguments to wordcloud()}
}
\value{
wordcloud
}
\description{
Function to make a wordcloud
}
\examples{
readLines("./data/compileText.txt") \%>\% str_replace_all("[[^[A-z] _]\\\\\\\\`]", " ") \%>\% str_split(" ") \%>\% unlist \%>\%  str_trim() \%>\% table() \%>\% as.data.frame(stringsAsFactors = F) \%>\% set_names(c("word", "freq")) \%>\% filter(nchar(word) > 0) -> tmp
tmp \%>\% MakeWordcloud()
}

