% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WordCloud.R
\name{MakeWordFreqCompare}
\alias{MakeWordFreqCompare}
\title{Create a word frequency comparison table from a list of words or phrases}
\usage{
MakeWordFreqCompare(wordlists, stem = T, rm.stopwords = T,
  stopword.list = NULL)
}
\arguments{
\item{wordlists}{list containing two vectors of strings}

\item{stem}{T/F - stem?}

\item{rm.stopwords}{T/F - remove english stopwords?}

\item{stopword.list}{List of stopwords to remove in addition to english stopwords (only matters if rm.stopwords is T)}
}
\value{
word frequency list
}
\description{
Create a word frequency comparison table from a list of words or phrases
}

