% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSolver.R
\name{mean_quantile_to_gamma_pars}
\alias{mean_quantile_to_gamma_pars}
\title{Function to find parameters for a gamma distn}
\usage{
mean_quantile_to_gamma_pars(mean, qu, p, init = c(1, 1), quiet = F, ...)
}
\arguments{
\item{mean}{expected value of the distribution}

\item{qu}{location (x) of p-th quantile.}

\item{p}{quantile (i.e. 100*p percent of the observed data is less than qu).}

\item{init}{starting values for optimization. Recommendation: (mean, 1), where x is close to the expected value (or median) of your distribution.}

\item{...}{Other options passed to optim}
}
\value{
shape and scale parameters for the gamma distribution, (a, s), where EX = a*s and Var(X) = a*s^2
}
\description{
Function to find parameters for a gamma distn
}

