% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_value.R
\name{na_to_value}
\alias{na_to_value}
\title{Function to convert all NAs in a vector to a specified value}
\usage{
na_to_value(vec, val = "")
}
\arguments{
\item{vec}{vector that potentially has NAs}

\item{val}{value to replace NAs with}
}
\value{
vector with no NAs
}
\description{
Function to convert all NAs in a vector to a specified value
}
\examples{
na_to_value(c(1:5, NA, 7:10), 6)
}

