% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ParameterSolver.R
\name{quantiles_to_gamma_pars}
\alias{quantiles_to_gamma_pars}
\title{Function to find parameters for a gamma distn}
\usage{
quantiles_to_gamma_pars(qu, p, init = c(1, 1), quiet = F, ...)
}
\arguments{
\item{qu}{location (x) of p-th quantile. Vector of length 2.}

\item{p}{quantile (i.e. 100*p percent of the observed data is less than qu). Vector of length 2.}

\item{init}{starting values for optimization. Recommendation: (x, 1), where x is close to the expected value (or median) of your distribution.}

\item{...}{Other options passed to optim}
}
\value{
shape and scale parameters for the gamma distribution, (a, s), where EX = a*s and Var(X) = a*s^2
}
\description{
Function to find parameters for a gamma distn
}

