% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WordCloud.R
\name{stemCompletion_mod}
\alias{stemCompletion_mod}
\title{Modify stem completion to go for the shortest word}
\usage{
stemCompletion_mod(x, dict, type = "shortest")
}
\arguments{
\item{x}{vector of strings}

\item{dict}{dictionary}

\item{type}{Defaults to "shortest"}
}
\value{
PlainTextDocument
}
\description{
Modify stem completion to go for the shortest word
}

