% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WordCloud.R
\name{wordcloud}
\alias{wordcloud}
\title{Wordcloud function, from the wordcloud package.}
\usage{
wordcloud(words, freq, scale = c(4, 0.5), min.freq = 3, max.words = Inf,
  random.order = TRUE, random.color = FALSE, rot.per = 0.1,
  colors = "black", ordered.colors = FALSE, use.r.layout = FALSE,
  fixed.asp = TRUE, asp = 1, ...)
}
\arguments{
\item{words}{the words}

\item{freq}{their frequencies}

\item{scale}{A vector of length 2 indicating the range of the size of the words.}

\item{min.freq}{words with frequency below min.freq will not be plotted}

\item{max.words}{Maximum number of words to be plotted. least frequent terms dropped}

\item{random.order}{plot words in random order. If false, they will be plotted in decreasing frequency}

\item{random.color}{choose colors randomly from the colors. If false, the color is chosen based on the frequency}

\item{rot.per}{proportion words with 90 degree rotation}

\item{colors}{color words from least to most frequent}

\item{ordered.colors}{if true, then colors are assigned to words in order}

\item{use.r.layout}{if false, then c++ code is used for collision detection, otherwise R is used}

\item{fixed.asp}{if TRUE the aspect ratio is fixed. Variable aspect ratio only supported if rot.per == 0}

\item{asp}{aspect ratio of plot (if fixed aspect ratio)}

\item{...}{Additional parameters to be passed to text (and strheight, strwidth).}
}
\description{
Wordcloud function, from the wordcloud package.
}

